Fix _vowelForMark map corruption and eliminate all compiler warnings
Engine.cpp — critical: _vowelForMark iterated with integer indices
The mark-key handler used `for (i=0; i<_vowelForMark.size(); i++)` and
accessed `_vowelForMark[i]` (integer keys 0..N). Because map::operator[]
inserts a default entry when the key is missing, each call polluted the
map with up to ~33 garbage empty entries for non-existent keys 0-38.
The loop only worked by accident — it eventually hit the real keys (KEY_E=26,
KEY_Y=29, KEY_U=30, KEY_I=31, KEY_O=32, KEY_A=38) as it iterated up.
Fixed by replacing integer-index loop with a proper map iterator loop.

Engine.cpp, ConvertTool.cpp, Macro.cpp — sign-compare warnings (31 total)
All `int` loop variable comparisons against `vector::size()` / `map::size()`
(which return `size_t`) now use `(int)` casts, eliminating every -Wsign-compare
warning. Build now produces zero warnings with -Wall on g++/C++11.

test_engine.cpp — expanded from 17 to 45 tests
New groups added:
- Mark on all vowel types (a/o/e/i/u/y) — directly validates the map fix
- _vowelForMark map integrity (10 repeated mark presses)
- All 5 Telex tone marks (s/f/r/x/j) on 'a'
- Code table switch (all 5 tables, aa->â)
- Caps-lock / shift flag preservation

+++ Sources/XoneKey/engine/ConvertTool.cpp
@@ -29,7 +29,7 @@ static vector<Uint8> _breakCode = {'.', '?', '!'};
 static bool findKeyCode(const Uint32& charCode, const Uint8& code, int& j, int& k) {
     //find character which has tone/mark
     for (map<Uint32, vector<Uint16>>::iterator it = _codeTable[code].begin(); it != _codeTable[code].end(); ++it) {
-        for (int z = 0; z < it->second.size(); z++) {
+        for (int z = 0; z < (int)it->second.size(); z++) {
             if (charCode == it->second[z]) {
                 j = it->first;
                 k = z;
@@ -61,10 +61,10 @@ string convertUtil(const string& sourceString) {
     if (convertToolToAllNonCaps)
         shouldUpperCase = false;
     
-    for (int i = 0; i < data.size(); i++) {
+    for (int i = 0; i < (int)data.size(); i++) {
         p = 0;
         //find char with tone/mark
-        if (i < data.size() - 1) {
+        if (i < (int)data.size() - 1) {
             switch (convertToolFromCode) {
                 case 2: //VNI
                 case 4: //1258
